Make x3 more useful error-prone rules errors instead of warnings
* http://errorprone.info/bugpattern/IntLongMath
* http://errorprone.info/bugpattern/FloatCast
* http://errorprone.info/bugpattern/ClassNewInstance
* http://errorprone.info/bugpattern/ConstructorInvokesOverridable
* http://errorprone.info/bugpattern/ThreeLetterTimeZoneID
* http://errorprone.info/bugpattern/UnsafeFinalization

and handle their impacts

+++ testutils/src/main/java/org/opendaylight/infrautils/testutils/concurrent/SlowExecutor.java
@@ -37,6 +37,6 @@ public class SlowExecutor implements Executor {
     }
 
     private long random(int low, int high) {
-        return random.nextInt(high - low) + low;
+        return (long) random.nextInt(high - low) + low;
     }
 }
